treewide: Remove IRQF_NO_SUSPEND
The qmp interrupt is a hardware wake capable interrupt. Using the
IRQF_NO_SUSPEND flag still allows this interrupt to resume the CPUs
because the interrupt is left unmasked on the gic. Remove this flag
and use enable_irq_wake so the wakeup interrupt is properly accounted
for.

* irqchip/qcom: mpm: Remove IRQF_NO_SUSPEND
Adithya R <gh0strider.2k18.reborn@gmail.com>

Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Ananth Raghavan Subramanian <sananth@codeaurora.org>
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>
Change-Id: I7e55b9feff43da6bda5d1eca12bf92f98be1c9f7
Bug: 123377615
Bug: 131260677
Bug: 143857220
Signed-off-by: Kyle Lin <kylelin@google.com>
Co-authored-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: pix106 <sbordenave@gmail.com>
Signed-off-by: Sã Śâjjãd <103353616+Sa-Sajjad@users.noreply.github.com>